requirements.txt
# Core scheduling and database dependencies
neo4j
networkx
python-dateutil
pytz
schedule

# Vector database and embeddings
chromadb
sentence-transformers

# CLI and configuration
click
pydantic
python-dotenv
typing-extensions

# OpenAI and async support
openai
aiohttp
uvloop

# Data processing utilities
pandas
numpy

# Development and testing dependencies
pytest
pytest-asyncio
pytest-cov
black
flake8
mypy

# Additional utilities
tqdm